fix broken formatters
NSUInteger is the same as uint32_t on iOS so there was an error thrown on iOS. Similarly for NSInteger. So instead I made formatters for int, long, unsigned int, and unsigned long which can be reused for NSInteger, NSUInteger, and uint32_t.

Also uint64_t and int64_t were incorrect. They have been changed to use "qu" and "qi" respectively.